pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "webspirit"
version = "0.0.3"
description = "A Python package for downloading resources from the web and organizing your bookmarks"
keywords = ["webspirit", "yt-dlp", "POO", "YouTube", "download", "bookmarks", "collections", "web-scrapping"]
readme = "README.md"
requires-python = ">=3.12.4"
license = {text = "MIT"}

authors = [
  { name = "Solange", email = "archange_paradise@proton.me" }
]

classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

dependencies = [
  "pandas==2.3.0",
  "colorlog==6.9.0",
  "yt-dlp==2025.6.9",
  "ipywidgets==8.1.7",
  "beautifulsoup4==4.13.5",

  "fastapi==0.127.1",
  "uvicorn==0.40.0",
  "pystray==0.19.5",
  "pillow==12.0.0"
]

[project.urls]
Homepage = "https://github.com/Archange-py/Webspirit/blob/main/README.md"
Repository = "https://github.com/Archange-py/Webspirit"
Library = "https://pypi.org/project/webspirit/"
Documentation = "https://archange-py.github.io/Webspirit"

[project.scripts]
webspirit = "webspirit.webspirit:main"

[project.optional-dependencies]
dev = ["pytest"]
